Bugfixes
Put Index.html and CSS file in same directory, keeping files in same directory allows the CSS to load in the HTML.  Corrected problems with glyphs, put a glyph, head, blurb, and index box for the icon well, centered the text, icons and input box.
Reworked Join form, spread things out a bit so it was a bit better on the eyes.  Seperated each part of the code into containers for clarity
Only thing that needs to be worked on is the fact the navbar is not responsive to the rest of the code.