More changes...
I fixed the dropdown navbar button, all it needed was class="navbar-inverse" in the nav declaration to take care of it.  Now, one issue that popped up was the fact the color on the navs was blending too much into the green background.  By making all of the navs set to class="active" they gained a black color with white text. so the navbar is mostly green, except for the navs, which are black.  This is the case for ALL views, collapsed or otherwise.  I didn't know right then how to change the colors of the active class to match, but I will look into it later to ensure everything is matching.  I also added a <a> tag for join because it did not have one.