the spacing for all of the components has been gone over.
The Jumbotronbox has a top and bottom margin to center it on the background, and the left/right margins were simply set to 0.5 em on all views.  I tried to keep a standard of 5rem spacing between components, and I was able to move the make an impact up by using a negative top margin.  As for the setup, I simply created <div classes> where needed, and defined them in CSS. Everything looks fairly decent now.  I still have not solved the problem with changing the navbar-nav colors to green and white (whether they are active or not, or whether they are a link or not).  That is my next goal.