navbar Green fixed
My big mistake was a typo.  To make the button for the collapsed bar come up, I needed to use navbar-inverse.  To fix the colors, I may have added a few things that can be simplified, but because I never specified navbar-inverse, it always assumed it was running under navbar-default, that was the problem.  I simply needed to ensure the navbar-inverse was declared in the CSS classes before anything would work correctly.