Features and Look Changed
All Icon-Well buttons are now based on <a href="#join" ....></a>.... "Make an Impact" was changed to the secondaryColor Instead of Green to break things up a bit.  The text under that has an extra sentence Telling the person to Ask Questions if they were in doubt about anything.  In the Dropbox of options, "Ask Questions!" was added as an option that could be selected.  There is JS code that could be used for a popup for the buttons in custom.js, I just need to figure out how to use it.